sattrack/ml-db/requirements.txt
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.26
python-dotenv>=1.0
supabase>=2.3
sgp4>=2.22
tenacity>=8.2
requests>=2.31
xgboost>=2.0
scikit-learn>=1.4
joblib>=1.3
